Another TODO: move unit from drive_t to diskunit_context_t.
Again this is really a deduplication, but one was signed an the other unsigned.
I settled on unsigned (like in drive_t) but a lot of printf formats needed
adjustment for that.


git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@44116 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/drive/iec/cia1571d.c
@@ -281,7 +281,7 @@ void cia1571_setup_context(diskunit_context_t *ctxptr)
 
     cia->prv = lib_malloc(sizeof(drivecia1571_context_t));
     cia1571p = (drivecia1571_context_t *)(cia->prv);
-    cia1571p->number = (unsigned int)(ctxptr->mynumber);
+    cia1571p->number = ctxptr->mynumber;
 
     cia->context = (void *)ctxptr;
 
@@ -294,7 +294,7 @@ void cia1571_setup_context(diskunit_context_t *ctxptr)
 
     cia->debugFlag = 0;
     cia->irq_line = IK_IRQ;
-    cia->myname = lib_msprintf("CIA1571D%d", ctxptr->mynumber);
+    cia->myname = lib_msprintf("CIA1571D%u", ctxptr->mynumber);
 
     cia1571p->diskunit = ctxptr;
